+++ src/Main.java
@@ -14,6 +14,7 @@ import ninja.mbedded.ninjaterm.util.loggerUtils.LoggerUtils;
 import ninja.mbedded.ninjaterm.util.stringUtils.StringUtils;
 import ninja.mbedded.ninjaterm.view.mainWindow.MainWindowViewController;
 import ninja.mbedded.ninjaterm.view.splashScreen.SplashScreenViewController;
+import org.apache.commons.cli.*;
 import org.controlsfx.glyphfont.GlyphFont;
 import org.slf4j.Logger;
 
@@ -31,6 +32,8 @@ public class Main extends Application {
     private Stage mainStage;
 
     private Logger logger = LoggerUtils.createLoggerFor(getClass().getName());
+
+    private Stage primaryStage;
 
     @Override
     public void start(Stage primaryStage) throws Exception {
@@ -54,14 +57,54 @@ public class Main extends Application {
         //======== COMMAND-LINE ARGUMENT PARSING =======//
         //==============================================//
 
-        logger.debug("Parsing command-line parameters. args = " + getParameters().getRaw());
-        for (String arg : getParameters().getRaw()) {
-            if (arg.equals("no-splash"))
-                disableSplashScreen = true;
+        Options options = new Options();
 
-            if (arg.equals("debug"))
-                LoggerUtils.startDebuggingToFile();
+        Option logc = Option.builder(null).longOpt("logc").numberOfArgs(0).desc("log to console").build();
+        options.addOption(logc);
+
+        Option logf = Option.builder(null).longOpt("logf").numberOfArgs(0).desc("log to file").build();
+        options.addOption(logf);
+
+        Option noSplash = Option.builder(null).longOpt("nosplash").numberOfArgs(0).desc("disable splashscreen").build();
+        options.addOption(noSplash);
+
+        CommandLineParser parser = new DefaultParser();
+        HelpFormatter formatter = new HelpFormatter();
+        CommandLine cmd;
+
+
+        try {
+            cmd = parser.parse(options, getParameters().getRaw().toArray(new String[0]));
+        } catch (ParseException e) {
+            System.out.println(e.getMessage());
+            formatter.printHelp("utility-name", options);
+
+            System.exit(1);
+            return;
         }
+
+//        logger.debug("Parsing command-line parameters. args = " + getParameters().getRaw());
+//        for (String arg : getParameters().getRaw()) {
+//            if (arg.equals("no-splash"))
+//                disableSplashScreen = true;
+//
+//            if (arg.equals("debug"))
+//                LoggerUtils.startDebuggingToFile();
+//        }
+
+        if(cmd.hasOption(logc.getLongOpt())) {
+            LoggerUtils.consoleLoggingEnabled.set(true);
+        }
+
+        if(cmd.hasOption(logf.getLongOpt())) {
+            LoggerUtils.fileLoggingEnabled.set(true);
+        }
+
+        if(cmd.hasOption(noSplash.getLongOpt())) {
+            disableSplashScreen = true;
+        }
+
+        this.primaryStage = primaryStage;
 
         if (disableSplashScreen) {
             // Skip this function, and go straight to loading the main window.
@@ -69,6 +112,11 @@ public class Main extends Application {
             return;
         }
 
+        loadSplashScreen();
+
+    }
+
+    private void loadSplashScreen() {
         this.splashScreenStage = primaryStage;
 
         // Load splashscreen FXML file and get controller